requirements.txt
flask

python-dotenv
WTForms
fairpyx
prtpy